requirements.txt
streamlit
tensorflow
numpy
pillow  # Remplace opencv-python par pillow